test: cover uniform list and tree placement parity

+++ rados/src/crush/decode.rs
@@ -295,7 +295,7 @@ fn decode_bucket(data: &mut Bytes, alg: u32) -> Result<CrushBucket> {
             }
         }
         BucketAlgorithm::Tree => {
-            let num_nodes = u32::decode(data, 0)?;
+            let num_nodes = u32::from(u8::decode(data, 0)?);
             if num_nodes > 10000 {
                 return Err(CrushError::DecodeError(format!(
                     "Tree num_nodes {num_nodes} exceeds maximum 10000"
